requirements.txt
requests>=2.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
python-dotenv>=1.0.0
openai>=1.3.0
torch>=2.0.0
transformers>=4.21.0
streamlit==1.45.1